requirements.txt
asgiref==3.4.1
Django==3.2.8
django-phonenumber-field==5.2.0
phonenumberslite==8.12.35
Pillow==8.4.0
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
